Mirror + contributing docs
Copying the
[Contributing](https://github.com/PataphysicalSociety/soupault#contributing)
block of Soupault who allows read/write contributions through the
proprietary GitHub platform or the German nonprofit ran FOSS Gitea
instance at Codeberg. This would allow users to make contributions and
have comments on a platform that is open. Eventually and sooner rather
than later, Gitea will be supporting federation so contributions won’t
require an account on either platform to contribute—which is better for
lock-in and privacy.