Work around problems with git-lfs
At least one package in the package set requires git-lfs, which causes
`fetchGit` to fail. In order to deal with this, we need run the nix
command that fetches these package in an environment with a particular
environment variable set. This commit sets that variable in a devshell.
Once the packages if fetched, it is cached and this is not needed.